+++ onnxruntime/core/common/cpuid_info.h
@@ -4,14 +4,7 @@
 #pragma once
 
 #include "core/common/common.h"
-
-#if defined(_M_IX86) || (defined(_M_X64) && !defined(_M_ARM64EC)) || defined(__i386__) || defined(__x86_64__)
-#define CPUIDINFO_ARCH_X86
-#endif
-
-#if defined(_M_ARM64) || defined(__aarch64__) || defined(_M_ARM) || defined(__arm__)
-#define CPUIDINFO_ARCH_ARM
-#endif  // ARM or ARM64
+#include "core/common/cpuid_arch_definition.h"
 
 namespace onnxruntime {
 
@@ -31,7 +24,7 @@ class CPUIDInfo {
   bool HasSSE4_1() const { return has_sse4_1_; }
   bool IsHybrid() const { return is_hybrid_; }
 
-  // ARM 
+  // ARM
   bool HasArmNeonDot() const { return has_arm_neon_dot_; }
 
   uint32_t GetCurrentCoreIdx() const;
@@ -72,7 +65,7 @@ class CPUIDInfo {
     }
     return is_armv8_narrow_ld_[coreId];
   }
-  
+
   /**
   * @brief Some ARMv8 power efficient core has narrower 64b load/store
   *        that needs specialized optimiztion in kernels
